Fixed One Note UI Issue and found another
so Now I have the UI scaling correctly BUT saving is NOT working at all, it seems like im working on two seperate files (the one on disk) and (the one in the editor) and the two only merge when i completly shut down the editor. I asked in the godot discord for help hopefully they know how to fix it because i sure as hell dont!